Change: Increase company CC storage to 32 bits to allow for custom colours.

+++ src/company_cmd.h
@@ -15,7 +15,7 @@
 #include "livery.h"
 
 enum ClientID : uint32_t;
-enum Colours : uint8_t;
+enum Colours : uint32_t;
 
 CommandCost CmdCompanyCtrl(DoCommandFlags flags, CompanyCtrlAction cca, CompanyID company_id, CompanyRemoveReason reason, ClientID client_id);
 CommandCost CmdCompanyAllowListCtrl(DoCommandFlags flags, CompanyAllowListCtrlAction action, const std::string &public_key);
